Refactor to avoid repetitive code on Blind SQLi implementations
