+++ lib/msf/core/exploit/sqli/mysqli/common.rb
@@ -73,7 +73,7 @@ module Msf::Exploit::SQLi::MySQLi
     #   @param database [String] the name of a database, or a function call, or nil for the current database
     #   @return [String] The character encoding, or nil if the database does not exist, or another error occurs
     #
-    def enum_database_encoding(database='database()')
+    def enum_database_encoding(database = 'database()')
       dump_table_fields('information_schema.schemata', %w[DEFAULT_CHARACTER_SET_NAME],
                         "SCHEMA_NAME=#{database.include?('(') ? database : "'" + database + "'"}").flatten[0]
     end
@@ -83,14 +83,14 @@ module Msf::Exploit::SQLi::MySQLi
     #   @param database [String] the name of a database, or a function call, or nil for the current database
     #   @return [Array] An array of Strings, the table names in the given database
     #
-    def enum_table_names(database='database()')
+    def enum_table_names(database = 'database()')
       dump_table_fields('information_schema.tables', %w[table_name],
                         "table_schema=#{database.include?('(') ? database : "'" + database + "'"}").flatten
     end
 
-    def enum_view_names(database='database()')
+    def enum_view_names(database = 'database()')
       dump_table_fields('information_schema.views', %w[table_name],
-                         "table_schema=#{database.include?('(') ? database : "'" + database + "'"}").flatten
+                        "table_schema=#{database.include?('(') ? database : "'" + database + "'"}").flatten
     end
 
     #
