+++ lib/msf/core/exploit/sqli/mysqli/time_based_blind.rb
@@ -1,36 +1,15 @@
+require 'msf/core/exploit/sqli/utils/time_based_blind'
+
 #
 #   Time-Based Blind SQL injection support for MySQL
 #
 class Msf::Exploit::SQLi::MySQLi::TimeBasedBlind < Msf::Exploit::SQLi::MySQLi::Common
-  def initialize(datastore, framework, user_output, opts = {}, &query_proc)
-    super
-  end
+  include ::Msf::Exploit::SQLi::TimeBasedBlindMixin
 
-  #
-  # Runs an SQL query, and returns its results (time-based blind technique)
-  #
-  # @param query [String] The SQL query to execute
-  # @param output_charset [Range] The range of characters to expect in the output, optional
-  # @return [String] The query result
-  #
-  def run_sql(query, output_charset: nil)
-    # TODO: detect latency and update sleepdelay manually?
-    if output_charset.is_a?(Range) && output_charset.count > 0
-      known_bits, bits_to_guess = get_bitmask(output_charset)
-    else
-      known_bits = 0
-      bits_to_guess = 8
-    end
-    vprint_status "{SQLi} Executing (#{query})"
-    if @hex_encode_strings
-      query = hex_encode_strings(query)
-      vprint_status "{SQLi} Encoded to (#{query})"
-    end
-    # first, get the length of the output
-    output_length = blind_detect_length(query, true)
-    vprint_status "{SQLi} Time-based injection: expecting output of length #{output_length}"
-    # now, get the output, of the given length
-    blind_dump_data(query, output_length, known_bits, bits_to_guess, true)
+  def initialize(datastore, framework, user_output, opts = {}, &query_proc)
+    @sleepdelay = datastore['SqliDelay'] || 1
+    @sleepfunc = 'sleep'
+    super
   end
 
   #
@@ -42,17 +21,5 @@ class Msf::Exploit::SQLi::MySQLi::TimeBasedBlind < Msf::Exploit::SQLi::MySQLi::C
     out_true = blind_request("if(1=1,sleep(#{datastore['SqliDelay']}),0)")
     out_false = blind_request("if(1=2,sleep(#{datastore['SqliDelay']}),0)")
     out_true && !out_false
-  end
-
-  private
-
-  #
-  # Performs one request, and does timing measurement, should leak one bit of information
-  #
-  def blind_request(query)
-    time = Time.now
-    @query_proc.call(query)
-    diff = Time.now - time
-    diff >= datastore['SqliDelay']
   end
 end
