+++ lib/msf/core/exploit/sqli/sqlitei/time_based_blind.rb
@@ -1,7 +1,11 @@
+require 'msf/core/exploit/sqli/utils/time_based_blind'
+
 #
 #   Time-Based Blind SQL injection support for SQLite
 #
 class Msf::Exploit::SQLi::SQLitei::TimeBasedBlind < Msf::Exploit::SQLi::SQLitei::Common
+  include ::Msf::Exploit::SQLi::TimeBasedBlindMixin
+
   HEAVYQUERY_DETECTION_SAMPLE = 10 # number of times to check if the block takes a delay of @sleepdelay
 
   def initialize(datastore, framework, user_output, opts = {}, &query_proc)
@@ -13,25 +17,6 @@ class Msf::Exploit::SQLi::SQLitei::TimeBasedBlind < Msf::Exploit::SQLi::SQLitei:
       detect_heavyquery_parameter
     end
     vprint_status "randomblob parameter: #{@heavyquery_parameter}"
-  end
-
-  def run_sql(query, output_charset = nil)
-    if output_charset.is_a?(Range) && output_charset.count > 0
-      known_bits, bits_to_guess = get_bitmask(output_charset)
-    else
-      known_bits = 0
-      bits_to_guess = 8
-    end
-    vprint_status "{SQLi} Executing (#{query})"
-    if @hex_encode_strings
-      query = hex_encode_strings(query)
-      vprint_status "{SQLi} Encoded to (#{query})"
-    end
-    # first, get the length of the output
-    output_length = blind_detect_length(query, true)
-    vprint_status "{SQLi} Time-based injection: expecting output of length #{output_length}"
-    # now, get the output, of the given length
-    blind_dump_data(query, output_length, known_bits, bits_to_guess, true)
   end
 
   #
@@ -48,13 +33,6 @@ class Msf::Exploit::SQLi::SQLitei::TimeBasedBlind < Msf::Exploit::SQLi::SQLitei:
 
   private
 
-  def blind_request(query)
-    time = Time.now
-    @query_proc.call(query)
-    diff = Time.now - time
-    diff >= @sleepdelay
-  end
-
   def detect_heavyquery_parameter
     @heavyquery_parameter = 10000000
     loop do
@@ -62,5 +40,6 @@ class Msf::Exploit::SQLi::SQLitei::TimeBasedBlind < Msf::Exploit::SQLi::SQLitei:
 
       @heavyquery_parameter *= 2
     end
+    @heavyquery_parameter = @heavyquery_parameter * 3 / 2 # for safety
   end
 end
